+++ src/DFT.h
@@ -4,4 +4,4 @@
 
 // Discrete Fourier Transform
 // DFT is computed just to half size of vector x
-std::vector<complex> DFT(std::vector<float> x);
+std::vector<complex> DFT(std::vector<double> x);
