+++ src/DTF.h
@@ -4,11 +4,11 @@
 // Discrete Transfer Function
 class DTF {
 private:
-	float b0, b1, b2, a1, a2;
+	double b0, b1, b2, a1, a2;
 	complex n0, n1, p0, p1;
 	void ComputeZeroesPoles();
 public:
-	DTF(float b0, float b1, float b2, float a1, float a2);
+	DTF(double b0, double b1, double b2, double a1, double a2);
 	void GetZeroesPoles(complex &n0, complex &n1, complex &p0, complex &p1);
 
 	// H(e^jw)
